+++ src/refactoring/components/Cart/index.ts
@@ -1,0 +1,4 @@
+import CartDetails from "./CartDetails";
+import ProductList from "./ProductList";
+
+export { CartDetails, ProductList };
